Clean-up the codegen module to have a very small API surface
