+++ modules/codegen/src/smithy4s/codegen/internals/CodegenRecord.scala
@@ -14,17 +14,17 @@
  *  limitations under the License.
  */
 
-package smithy4s.codegen
+package smithy4s.codegen.internals
 
 import software.amazon.smithy.model.node.Node
 import scala.jdk.CollectionConverters._
 import software.amazon.smithy.model.Model
 
-private[smithy4s] final case class CodegenRecord(
+private[internals] final case class CodegenRecord(
     namespaces: List[String]
 )
 
-private[smithy4s] object CodegenRecord {
+private[internals] object CodegenRecord {
 
   val METADATA_KEY = "smithy4sGenerated"
