+++ modules/codegen/src/smithy4s/codegen/internals/EnumUtil.scala
@@ -14,9 +14,20 @@
  *  limitations under the License.
  */
 
-package smithy4s.codegen
+package smithy4s.codegen.internals
 
-object EnumUtil {
+private[codegen] object EnumUtil {
+  def enumValueClassName(
+      name: Option[String],
+      value: String,
+      intValue: Int
+  ) = {
+    name.getOrElse {
+      val camel = toCamelCase(value).capitalize
+      if (camel.nonEmpty) camel else "Value" + intValue
+    }
+  }
+
   private def toCamelCase(value: String): String = {
     val (_, output) = value.foldLeft((false, "")) {
       case ((wasLastSkipped, str), c) =>
@@ -29,18 +40,6 @@ object EnumUtil {
         }
     }
     output
-  }
-
-  def enumValueClassName(
-      name: Option[String],
-      value: String,
-      intValue: Int
-  ) = {
-    name.getOrElse {
-      val camel = toCamelCase(value).capitalize
-      if (camel.nonEmpty) camel else "Value" + intValue
-    }
-
   }
 
 }
