+++ modules/codegen/src/smithy4s/codegen/internals/LineSyntax.scala
@@ -14,12 +14,12 @@
  *  limitations under the License.
  */
 
-package smithy4s.codegen
+package smithy4s.codegen.internals
 
 import cats.implicits._
-import smithy4s.codegen.WithValue.ToLineWithValue
+import WithValue.ToLineWithValue
 
-object LineSyntax {
+private[codegen] object LineSyntax {
   implicit class LineInterpolator(val sc: StringContext) extends AnyVal {
     def line(renderables: ToLineWithValue[_]*): Line = {
       renderAndCombine(renderables.toList)
