+++ modules/codegen/src/smithy4s/codegen/internals/ShapeExtractor.scala
