+++ modules/codegen/src/smithy4s/codegen/internals/SmithyResources.scala
