+++ modules/codegen/test/src/smithy4s/codegen/internals/SmithyToIRSpec.scala
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package smithy4s.codegen
+package smithy4s.codegen.internals
 
 import munit._
